backend/requirements.txt
# FastAPI и веб сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# FreeSWITCH интеграция
ESL==1.0.1
greenswitch==2.0.0

# AI сервисы
openai==1.3.5
elevenlabs==0.2.26
langchain==0.0.339

# Аудио обработка
pydub==0.25.1
soundfile==0.12.1
numpy==1.26.2
scipy==1.11.4

# Redis (опционально)
redis==5.0.1
aioredis==2.0.1

# Утилиты
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Мониторинг и логирование
loguru==0.7.2
sentry-sdk==1.38.0

# HTTP клиент
httpx==0.25.2
aiohttp==3.9.1

# Валидация
email-validator==2.1.0
phonenumbers==8.13.26

